consensus: move pkg to internal
