+++ cmd/tendermint/commands/replay.go
@@ -3,7 +3,7 @@ package commands
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/tendermint/tendermint/consensus"
+	"github.com/tendermint/tendermint/internal/consensus"
 )
 
 // ReplayCmd allows replaying of messages from the WAL.
